Cleanup

+++ database/migrations/2017_11_13_211529_create_purchases_table.php
@@ -18,7 +18,7 @@ class CreatePurchasesTable extends Migration
             $table->unsignedInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('description');
-            $table->decimal('amount');
+            $table->bigInteger('amount');
             $table->timestamps();
         });
     }
